Addressing some review comments
- Added new properties "Suffix" and "CameraIndex" to the Miniscope GUI configuration
- Changed the data type of saturation threshold from double to int for consistency.
- Updated the StatusBar to display average frame rate and frame number, while removing the buffer size display.
- Moved BufferSize to the time-series plots

+++ OpenEphys.Miniscope.Design.Gui/SaturationSettingsDto.cs
@@ -11,20 +11,20 @@ namespace OpenEphys.Miniscope.Design.Gui;
 /// </summary>
 /// <param name="Threshold">The minimum pixel intensity considered saturated.</param>
 /// <param name="Color">The color used to highlight saturated pixels.</param>
-public record SaturationSettingsDto(double Threshold, Scalar Color);
+public record SaturationSettingsDto(int Threshold, Scalar Color);
 
 /// <summary>
 /// Combines individual saturation setting values into a single <see cref="SaturationSettingsDto"/>.
 /// </summary>
 [Description("Combines individual saturation setting values into a single object.")]
-public class CreateSaturationSettingsDto : Transform<Tuple<double, Scalar>, SaturationSettingsDto>
+public class CreateSaturationSettingsDto : Transform<Tuple<int, Scalar>, SaturationSettingsDto>
 {
     /// <summary>
     /// Creates a <see cref="SaturationSettingsDto"/> from a sequence of tuples containing the saturation settings.
     /// </summary>
     /// <param name="source">A sequence of tuples containing the threshold and color values.</param>
     /// <returns>A sequence of <see cref="SaturationSettingsDto"/> objects.</returns>
-    public override IObservable<SaturationSettingsDto> Process(IObservable<Tuple<double, Scalar>> source)
+    public override IObservable<SaturationSettingsDto> Process(IObservable<Tuple<int, Scalar>> source)
     {
         return source.Select(value => new SaturationSettingsDto(value.Item1, value.Item2));
     }
